Reduced Shader to fewer slots
The shader in glTFast takes a lot of memory and resources due to a large
amount of variants. This conflicts with our WebGL builds and as a
quickfix we have replaced the shader with a more minimal version.

In the future this could be more elegantly solved using Dynamic Shader
Variant loading, but this was not available at time of writing.

Original author: Sam Baas